Update rpmanager.sh
This script is intended to use when RetroPie-Manager is installed via retropie_setup.sh. It happens because of the hardcoded path in the rpmanager_dir variable.

The RetroPie-Manager can not be started directly by the root user, but if it's called by a "sudo environment", it's OK. The sudo user will start the service.

The default TCP port is 8000, but the user can define another port using the --port or -p option (allowed ports are between 1024 and 65535, inclusive).  This script doesn't allow to start multiple instances of RetroPie-Manager, even if the user try to start it listening to a different port.

If the --log option is used, the log messages will be saved in "/opt/retropie/supplementary/retropie-manager/logs" ($rpmanager_dir/logs) with an appropriate file name

Execute it with --help to see the available options.