there was a missing '!' at line 250
This script MUST be in the same directory RetroPie-Manager's directory.  If it's installed in "/opt/retropie/supplementary/retropie-manager/", and only in this case, the user can make symbolic links to it. Otherwise, the symbolic links won't work.

The RetroPie-Manager can not be started directly by the root user, but but if you need it (ex.: on boot), use the --user option. The user must be a RetroPie user (must have a RetroPie directory tree in its homedir).  If it's called in a "sudo environment", it's OK, the sudo user will start the service.

The default TCP port is 8000, but the user can define another port using the --port or -p option (allowed ports are between 1024 and 65535, inclusive).  This script doesn't allow to start multiple instances of RetroPie-Manager, even if the user try to start it listening to a different port.  

If the --log option is used, the log messages will be saved in "/opt/retropie/supplementary/retropie-manager/logs" ($rpmanager_dir/logs) with an appropriate file name.

Execute it with --help to see the available options.